+++ src/ui/docking_tool_window.cc
@@ -24,8 +24,7 @@ DockingToolWindow::~DockingToolWindow() {
 }
 
 Rect DockingToolWindow::RectForTitleBar() {
-  return Rect(
-      0, 0, Width(), static_cast<int>(Skin::current().title_bar_size()));
+  return Rect(0, 0, Width(), Skin::current().title_bar_size());
 }
 
 void DockingToolWindow::Render() {
